:zap: Attemot to improve batching further
Whilst everything I have done here makes sense to me, and whilst it seems to be able to handle spikes in activity very well, there is a bit of a discrepency in performance. It can peak to reach 25000 inserts per second, and then spend tens of minutes idling around 50 inserts/second (on either links or pages). I don't know why this is at all??